dev: change backend: go

+++ cmd/root.go
@@ -1,12 +1,9 @@
 package cmd
 
 import (
-	"bytes"
-	"fmt"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 	"os"
-	"os/exec"
 	"path/filepath"
 )
 
@@ -21,15 +18,7 @@ var rootCmd = &cobra.Command{
 	PersistentPreRunE: func(_ *cobra.Command, _ []string) error {
 		if GlobalOptions.Password == "" {
 			// 通过 openssl 生成密码
-			c := exec.Command("openssl", "rand", "-base64", "16")
-			var out bytes.Buffer
-			c.Stdout = &out
-			err := c.Run()
-			if err != nil {
-				return err
-			}
-			GlobalOptions.Password = out.String()
-			fmt.Println("passwd:", GlobalOptions.Password)
+			GlobalOptions.Password = "ksp123456"
 		}
 		return nil
 	},
@@ -44,23 +33,27 @@ func init() {
 	rootCmd.PersistentFlags().StringVarP(&GlobalOptions.User, "user", "u", "admin", "user")
 	rootCmd.PersistentFlags().StringVarP(&GlobalOptions.Password, "password", "p", "", "passwd")
 	rootCmd.PersistentFlags().StringVarP(&GlobalOptions.FilePath, "file", "f", "", "file path")
+	rootCmd.PersistentFlags().StringVarP(&GlobalOptions.Dir, "dir", "d", "", "dir path")
 	rootCmd.RunE = func(cmd *cobra.Command, args []string) error {
 		user := GlobalOptions.User
 		passwd := GlobalOptions.Password
 		filePath := GlobalOptions.FilePath
-		_, err := os.Stat(filePath)
-		if err == nil {
-			err = os.Remove(filePath)
-			if err != nil {
-				return errors.Errorf("failed to remove the binary file: %s, err: %s", filePath, err.Error())
+		dir := &GlobalOptions.Dir
+		if *dir != "" {
+			filePath = filepath.Join(*dir, ".htpasswd")
+		} else {
+			_, err := os.Stat(filePath)
+			if err == nil {
+				err = os.Remove(filePath)
+				if err != nil {
+					return errors.Errorf("failed to remove the binary file: %s, err: %s", filePath, err.Error())
+				}
 			}
-		}
-
-		fileDir := filepath.Dir(filePath)
-
-		err = os.MkdirAll(fileDir, os.ModePerm)
-		if err != nil {
-			return errors.Errorf("failed to create directory: %s, err: %s", fileDir, err.Error())
+			fileDir := filepath.Dir(filePath)
+			err = os.MkdirAll(fileDir, os.ModePerm)
+			if err != nil {
+				return errors.Errorf("failed to create directory: %s, err: %s", fileDir, err.Error())
+			}
 		}
 
 		f, err := os.Create(filePath)
